Do not build a static library anymore
This does not work on Linux systems anyway as libudev-dev does not
provide a static library, and providing one for non-Linux systems would
make handling this inconsistent in depending packages.

In addition newer glibc generate a link-time warning on code using
getpwuid() calls, requesting to use the dynamic library version.